requirements.txt.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
bidict==0.23.1
biosppy==2.1.2
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.49.0
future==1.0.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.0
h5py==3.10.0
heartpy==1.2.7
idna==3.6
ipykernel==6.29.3
ipython==8.22.1
iso8601==2.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
nest-asyncio==1.6.0
neurokit2==0.2.7
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyemd==1.0.0
Pygments==2.17.2
pyhht==0.1.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.5.0
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
shortuuid==1.0.11
six==1.16.0
SQLAlchemy==2.0.27
stack-data==0.6.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.22.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tftb==0.1.4
threadpoolctl==3.3.0
tornado==6.4
traitlets==5.14.1
typeguard==2.13.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
WTForms==3.1.2